Move app related methods to separate source file
